fix:dependencia ciclica e recuperar senha
