+++ src/navegation.js
@@ -1,0 +1,5 @@
+export const navigateTo = (url) => {
+  window.history.pushState({}, '', url);
+  const popstateEvent = new PopStateEvent('popstate', { state: {} });
+  dispatchEvent(popstateEvent);
+};
