--- src/pages/feed/confirm.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line
-export const confirmAction = (msg) => confirm(msg);
