+++ src/services/confirm.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line
+export const confirmAction = (msg) => confirm(msg);
