+++ src/style.css
@@ -5,7 +5,7 @@
 }
 
 .background{
-    background: linear-gradient(to bottom, #0f0c29, #302b63, #24243e) ;
+    background: linear-gradient(to bottom, #0f0c29, #302b63, #24243e) ;   
 }
 
 .container-main{
@@ -109,7 +109,7 @@ input {
 }
 
 .login{
-    height: 460px;
+    height: 600px;
     background: #eee;
     border-radius: 60% / 10%;
     transform: translateY(-180px);
